Una sola puerta en la base que se entrega, y arreglar la pantalla que la abre
La siembra creaba once usuarios y diez tenían la clave escrita en un repositorio
público. Ahora una base nueva se entrega con uno solo, direccionia, y desde la
pantalla Usuarios se crean la cuenta de coordinación y los usuarios de cada
punto. Los diez de desarrollo pasan al bloque de ejemplos, que nunca llega a un
Postgres de verdad. `npm run db:usuarios` escribe el SQL que deja así una base ya
creada: borra a los que nunca cargaron nada y desactiva a los que sí, porque
borrarlos rompería la trazabilidad.

Para que eso sea posible hubo que arreglar la pantalla, que no servía:

  · No se podía crear una cuenta de coordinación: el alta escribía 'vigilador'
    fijo y exigía un punto. La pantalla lo decía sin vueltas —«El de coordinación
    se crea en la base»—, que es exactamente lo que no puede pasar en algo que se
    entrega.

  · No se podía cambiar la contraseña de una cuenta de coordinación desde ningún
    lado. El README aseguraba dos veces que sí.

  · NINGUNA acción de fila funcionaba: ni resetear PIN, ni activar, ni
    desactivar, ni desbloquear. Eran varios botones `name="accion"` en un solo
    formulario —lo natural en HTML, el navegador manda el valor del botón que
    envió— pero React no serializa el botón que envía cuando la acción del
    formulario es una server action: al servidor llegaba `id` y nada más. Ahora
    cada acción tiene su propio formulario y lleva la suya en un campo oculto.

  · El camino de la credencial equivocada reventaba con un error de servidor en
    vez de decir «no es correcta». Postgres deducía dos tipos distintos para el
    mismo parámetro —«inconsistent types deduced for parameter $2, text versus
    integer»— y el UPDATE de intentos fallidos no se podía ni preparar. Le falta
    a cualquier vigilador que erre el PIN una vez.

La contraseña de coordinación va escrita y de doce caracteres para arriba; el
sistema no la inventa ni la muestra. El PIN de punto sigue siendo de cuatro
dígitos y lo puede inventar: se teclea en la calle, y lo que lo protege es el
bloqueo por intentos y que no pueda escribir fuera de su punto.

Y lo que se ve cuando todavía no hay usuarios de punto: la pantalla de ingreso lo
dice y ofrece sólo el acceso de coordinación, en vez de un selector vacío.

db:verificar ya no depende de los usuarios sembrados: se crea los cuatro que
necesita, desactivados para que no asomen en el selector de ingreso, y los borra
al final. Verificado en las tres formas —base de producción con un usuario 49/49,
base con ejemplos 50/50, y repetido— sobre un PGlite con los permisos de fábrica
de Supabase puestos antes de crear las tablas.

+++ src/lib/acceso.ts
@@ -62,9 +62,16 @@ export async function verificarAcceso(usuario: string, credencial: string): Prom
     if (!coincide) {
       const intentos = p.intentos_fallidos + 1
       await tx.consultar(
+        // Los tipos van escritos. Sin los casts, Postgres tiene que deducir el
+        // de $2 dos veces —en la asignación y en la comparación— y saca dos
+        // distintos: «inconsistent types deduced for parameter $2, text versus
+        // integer». Reventaba el ingreso entero, y sólo en el camino de la
+        // credencial equivocada, que es el que nadie prueba.
         `update perfiles
-            set intentos_fallidos = $2,
-                bloqueado_hasta = case when $2 >= $3 then now() + make_interval(mins => $4) else null end
+            set intentos_fallidos = $2::int,
+                bloqueado_hasta = case when $2::int >= $3::int
+                                       then now() + make_interval(mins => $4::int)
+                                       else null end
           where id = $1`,
         [p.id, intentos, INTENTOS_MAXIMOS, MINUTOS_BLOQUEO],
       )
